+++ lib/corona/db/db.rb
@@ -3,8 +3,8 @@ module Corona
     require 'sequel'
     require 'sqlite3'
     def initialize
-      Sequel::Model.plugin :schema
       @db = Sequel.sqlite(CFG.db, :max_connections => 1, :pool_timeout => 60)
+      create_table #unless @db.table_exists?(:devices)
       require_relative 'model'
     end
 
@@ -27,5 +27,25 @@ module Corona
       sysName  = Device[:oid_sysName => oid_sysName]
       [ip, sysName]
     end
+
+    private
+    
+    def create_table
+      @db.create_table? :devices do
+        primary_key :id
+        String      :ip
+        String      :ptr
+        String      :model
+        String      :oid_ifDescr
+        Boolean     :active
+        Time        :first_seen
+        Time        :last_seen
+        String      :oid_sysName
+        String      :oid_sysLocation
+        String      :oid_sysDescr
+        String      :oid_sysObjectID
+      end
+    end
+
   end
 end